Move models to a separate folder
